Fix duplicate builtin classlike member indexing
Reflection also returns the members of parents as if they were members
of the child class. As we already do our own inheritance and interface
resolution for user-defined classes, we can do this ourselves.
Furthermore, it is not only incorrect in our setup, but methods that
aren't directly part of the class end up in the database, take
up space, and decrease performance.