Unset the cli.pager option within the main script
Having this option set to a paging application such as less causes the
output received by the CoffeeScript side of the API to be paged, which
it perceives as duplication. The option provides benefits in userland,
but when the CLI is being used programatically there is no need for it,
so we can safely override it.

Fixes issue php-integrator/atom-base#189