Remove shared locking code
As there may be multiple server processes at the same time, sharing
a lock file may cause parallel server processes to hang up whilst the
other is accessing the file.

The shared lock file was originally placed because multiple processes
operating on the same project and database may be active simultaneously,
but in a socket server scenario there should only ever be one socket
server working with the database and cache of the project. Other socket
servers will be working on another database (and different cache folder)
entirely.