Reenable progress streaming
It turns out that this was working as intended, but the data was never
delivered to the socket. This happened because reindexing happens inside
its own loop, which causes control to never end up back in the React
event loop in the meantime (which handles the actual writing).

The data was simply being buffered in the stream's buffer until the
reindexing loop finished and the next tick in React's main loop wrote
it to the socket.