Start experimenting with an autocorrecting (partial) parser
This adds an experimental "autocorrecting" parser that attempts to
correct PHP expressions that are invalid according to PhpParser.
It will automatically attempt to correct the expression by inserting
things such as trailing semicolons or a bogus property name (useful
for unfinished expressions suchas "self::").

We currently still have a hack where we use the PartialParser to
extract a "sanitized call stack", which is basically a simplified
representation of the parts of an expression in array format, such as
['$this', 'someMethod()', 'foo']. Because of this, we have to bend
ourselves backwards in the TypeDeducer to parse this string format.

To make things worse, because of legacy from the CoffeeScript side, the
TypeDeducer has this string parsing as its main way of deducing types
and needs to convert any (PhpParser) node to said string representation
before being able to deduce its type.

Life would be much easier if we could use PhpParser nodes everywhere
and if TypeDeducer just worked with nodes.

This work is still very much experimental and WIP. It may turn out to be
impossible to implement this and retain all current functionality.