Include "class" constant in class constant data
The "class" constant, which was introduced in PHP 5.5, is available for
all classes and always resolves to its FQCN. It is now returned
transparantly along with the rest of the class constants.

This also fixes the issue where the unknown member linter kept
complaining that it didn't know any member called "class".