Enforce libraries compatible with PHP 5.6
This is not really a problem for users of the core, but as the core
is used directly by Atom (and thus by users using all kinds of PHP
versions without them installing the dependencies themselves), we have
to ensure we keep to the lowest supported version.

When we manage to remove the Composer dependencies from Git, we'll be
able to remove this clause again.