Throw errors when expression is empty
The expression should never be empty if the code isn't or else we
failed fetching the expression for some reason. For invocation info
this is alright as there may be cases with incomplete PHP code that we
can't deal with. For other uses, such as parsing the default value
of properties and constants, it is not acceptable and an exception
should be thrown.