Don't do a nested transaction
This increased performance a little bit as the file indexer does a
transaction by itself, which is then ignored because one is already
active. However, when an exception occurred inside the FileIndexer that
was not caught by the ProjectIndexer, the transaction was not rolled
back nor committed and left in an unfinished state, causing problems
when the next one was started and committed or rolled back.

References https://github.com/php-integrator/atom-base/issues/256