Skip local types for self, static and $this
When the type of a local expression resolves to self, static or $this,
it should not subsequently make those types be subject to more local
types. In other words, the type of self should always be the type of the
class the type was in (which can never be different, not even by somehow
changing that by using e.g. conditionals or type overrides).