Require at least PHP 7.1
Updating to PHP 7 is neat, but adding type hints across the entire
code base for PHP 7 is limited to those supported by PHP 7, i.e.
void return types and nullable types are not supported yet. This
means that one has to walk through the entire codebase *again* once
the upgrade to PHP 7.1 happens, which is a bit too much to ask for the
sole gain of maintaining compatibility with a PHP release that has
already been superceded by a newer one and whose active support ceases
by the end of the year, which may not be far from when 3.0 is actually
released.

Since this may impact Atom end users, a means to maintain the current
packages as is will be investigated. For example, it might be possible
that the current Atom packages will automatically update to a new core,
say 3.0, and require PHP 7.1, whilst a fork of the packages of the last
known version to support PHP 5.6 and PHP 7.0 will be made and published
to the Atom package registry as something like
php-integrator-base-legacy-5.6, php-integrator-tooltips-legacy-5.6,
and so forth.