Extract LastExpressionParser from PartialParser
This also causes the indexers to not attempt to extract the last part of
the expression. Instead, they will only use the modified partial parser,
which now only attempts to make corrections in order to be able to pars
the code, but is not capable of parsing multiple expressions and only
fetching the last value. This should however not be necessary, as this
parser is only used during indexing to determine the type of default
values for properties and constants, which are always a single
expression

References #83