Fix nested function call corner cases
With nested function calls, the signature help of the parent function
was being fetched when inside the parantheses of the nested function
call when it had no arguments.

This should also fix an issue when the position was just before the
closing paranthesis of a nested function call, where the parent
signature help would be displayed instead of the nested function
signature help.