Divide test suite in two
This also causes unit tests to run before integration tests. This is
more interesting than the other way around (or mixing them), as unit
tests test for specific errors and corner cases on a detailed level.
Integration tests test the larger picture (the integration) and will
likely trip somewhere on the same problem as the unit tests, but the
error message will be much less specific and useful for the developer.