Remove isClassType check
Except self, parent, and static, there are no reserved PHP keywords that
will end up inside a Node\Name and still be valid. For example, this
change will cause the following to be picked up as well:

    int::foo();

This will now complain that classlike "int" was not found, where
previously it did not, because "int" is not a class type.

I don't really see any of this as an issue, however, since there is no
way this code is semantically valid and adding code to remove one more
error message about it isn't worth the hassle. Also, the error message
isn't wrong, as you're trying to use "int" as a classlike.