Move attaching node metadata to decorator
This moves node metadata to a parser decorator. Doing traversing after
parsing may be expensive if the metadata isn't used. However, quite a
few parts of the codebase are already dependent on this metadata. As a
result, these were all traversing the nodes to attach this metadata
already. As a result, the same logic was getting duplicated in several
locations and needed to be duplicated over and over again.

Even though it's not that good for performance to traverse the tree and
attach the metadata when it isn't necessary, the traversing may now only
happen once if there are multiple queries running over the same segment
of code. This happens because the MetadataAttachingParser is now also
decorated by the CachingParserProxy, which will check its cache to see
if the source was already parsed.