Don't parse default values a second time
We already have this information from the outline itself, there is no
use in parsing it twice, which is bad for performance.

Additionally, this fixes an issue where structural elements with a
function call as default value don't trip the
FunctionNameNodeFqsenDeterminer because they the result of the default
value parsing doesn't have any namespace attached.