Replace BuiltinIndexer with PhpStorm stub indexing
There are still some HTML tags and stuff in the docblocks of the stubs
that we need to clean up in the future. For the moment, however, this
beats having to rewrite the entire BuiltinIndexer into Doctrine. This
also takes away quite a maintenance burden of maintaining separate
indexers and improves the built-in suggestions.

This doesn't include any version checks with the tags from the stubs
yet.

Fixes #2