Store FQCN's in trait aliases and precedences
This happens for the same reason as the reason structures now reference
FQCN's; to ensure that associations aren't lost when the referenced
item is deleted. As when it is readded (e.g. in another file), we can't
afford to traverse the entire class hierarchy again to see where we need
to readd the association that was removed.