Fix closing connection not reinitializing entity manager
On startup, the client first calls "test" to check the databaes, which
created the initial connection and entity manager. Afterwards, as the
database did not yet exist, it would return false, causing the client to
initiate an "initialize command".

The initialize command would ensure the connection is closed to remove
the new database that had been created by the test command, but it would
not create a new entity manager.

For some very strange reason, the PRAGMA's in SqliteConnectionFactory
were being ignored because of this (tested by increasing the WAL
checkpoint size there), even though the connection was created and
these PRAGMA's were issued in both situations.

By increasing the WAL checkpoint size in SqliteConnectionFactory, it was
clear that the PRAGMA's were being ignored as the size of the WAL file
would never rise above the default (to the increased limit), whilst it
does after this fix.

The other PRAGMA's were also being ignored, resulting in a major
slowdown in indexing performance (taking several minutes to index
PHP stubs whilst it now only takes about 5 seconds).