Run all tests, even if errors occurred
I changed this before, but now realized that it is only inconvenient
when using the command line. With an IDE or visual testing tool (like
PhpStorm, the tester package for Atom, ...), seeing all errors is more
useful as the first error encountered is not always the one that is
actually exposing the root cause.