Fix CachingParser pretty much never caching
This happened because the error handler's object hash is taken up into
the cache key. Pretty much every caller specifies its own collecting
error handler, which is a different object, so they were all missing
the cache.

This also changes the CachingParser to not cache if something else is
used than a collecting parser, since behavior would change if you parse
a second time with a throwing error handler (i.e. no errors would be
thrown because the cache is hit). It would be possible to try and cache
the thrown exception, but it would need to be cloned and most of the
code base is using collecting handlers for error tolerance anyways.