Treat uncaught exceptions during file indexing as logic exceptions
Exceptions should not be propagating upwards during file indexing. If
there is an acceptable error, it should be caught in the indexing
visitors and handled appropriately.

Unacceptable exceptions shouldn't be occurring during indexing (save for
when PHP parser can't parse a file because it is too malformed) and they
indicate bugs that should be solved.