Fix (default) values being passed instead of code of file
This fixes an error that occurred specifically when a define contained
a ternary expression, which in turn contained a constant fetch
expression. On top of this, the offset of the constant fetch node had
to be larger than the length of the value of the define that was being
indexed.

This happened because the default value was being passed as code instead
of the actual code of the file during indexing.