Fix class relationships not being properly persisted on index
The object itself was properly updated, but the last change was not
persisted to storage because these changes happened after the initial
persist call.

This didn't get picked up by the tests because there was a bug in the
tests, where another variation of the test (where the entity manager was
cleared) did not get executed.