Utilize simple request queue
Instead of immediately handling a request and writing its response after
encountering it, requests are now sent to a queue and scheduled to be
handled as soon as possible.

This will allow the socket to drain (empty) before all its contained
requests are handled properly. As a side-effect, it may be possible to
cancel requests that haven't been handled yet in the future.

References #123