Fix progress events not being sent immediately
Apparantly reactphp's event loop will not send any actual writes or
finish an event loop iteration until all "nextTick" callbacks currently
in the queue have been handled. Because the ticks reschedule themselves
as long as our request queue is not empty, the entire processing of our
request queue is seen as "one tick" for the event loop. This is not
acceptable as we need actual responses to be sent whilst we continue
processing so indexing progress events can be streamed.

This replaces the tick system with a periodic timer that will have the
desired behavior.

The timer is uninstalled when the queue is empty so it will not continue
firing when the queue is empty (and thus hog the CPU).

References #123