Replace $this->assertX with static::assertX
This is mostly preference, but I'm selecting the static variant over the
non-static one because the methods are actually static themselves.

PHP supports both syntaxes, but calling static methods in a non-static
way does not clearly convey to the reader that the method is, in fact,
static.

Also, which is not personally relevant, but may be for other developers,
PhpStorm now warns about this practice.