Only prune on folder index requests
In a quick test scenario, a file reindex request took around 70 ms and
this pruning added 30 ms on top of that every time. This is a lot of
time for a scenario where only very rarely files were removed from disk.

Also, when reindexing a file, you usually don't care about other files
that may have been removed.

By only doing this when folders are present, vacuuming is limited to
folder reindex requests (e.g. when the project is reindexed), which
seems appropriate. In these scenarios, an extra 30 ms on top of the
time needed for these requests is more appropriate.

References #139