Make most classes final
This is somewhat of a code BC break, though personally I think this is
an acceptable one as the core was never meant to be used as a library
in other projects anyway. It is not a BC break that will affect
consumers (clients) of the core.

At some point in the future, more of the core may be split off into
actual libraries (some of it already is).