Fix "Position out of bounds" with modified, unindexed, file contents
This hopefully fixes #126. It should if the problem was what I described
in that issue. I'll have to take it for a test drive for a while to see
if it is indeed fixed.

In any case, this is an improvement regardless as this will allow for
request priorities in the future. Also, a client should not have to know
when the contents of a new request are different from the previous one
so that it should first send a reindex request for them.

Fixes #126