Fix large performance hit when reindexing namespaces and imports
Cleaning up old namespaces and imports in a file during reindexing was
taking a very long time. It turned out that this was not due to indexing
or persistence itself, but due to the event listener that was in turn
updating the namespace list registry.

The namespace list registry was using array_search and thus had O(n)
performance. If this list was large (and there are two namespace entries
for each file that has a namespace at the top, multiplied by the
amount of namespaced files in the code base), this lookup time could
severely slow down removing the old namespaces from the list and adding
the new ones.

A quick measurement with a file of 200 lines and 2 such namespaces
indicated it took over 100 ms to reindex, of which around ~ 50 ms were
taken for these kind of updates. This difference is huge in commands
that are supposed to respond in a timely manner, such as type deduction
and signature help.

This issue became especially noticable since #126 caused indexing to be
performed on new code before any request is processed.

References #139