Add class to check if function autocompletion applies in specified context
These tests probably won't cover all the correct exclusion and inclusion
cases, but we can always finetune later. In any case, it will always be
better than what we did with Atom's autocompletion, since that was just
a rudimentary regex check that only stopped (global) function
autocompletion if it detected a :: or -> in front of hte cursor.

References #42