Add simple fuzzy matching autocompletion provider
Here's to hoping that this will be a decent implementation. I'd rather
not implement my own fuzzy matching, but the language server protocol
doesn't describe any ability for clients to do this for the server.
Also, having this wouldn't solve the problem where we're sending back
massive lists of suggestions (class, function and constant lists) over
the socket to have them filtered by the client each time. So, the only
thing we can do is fuzzy match server-side to limit the results as much
as possible before sending them back.

Fuzzy matching will also allow us to sort the most interesting matches
up top as well as limit the total suggestion count to a sane number (
e.g. 15 or so) later on, which further reduces the data that needs to
be sent to the client.

References #42