Replace LogicException with AssertionError
AssertionError is the direct result of an assert(), which is what we'd
been using if it weren't automatically disabled in production mode
since we want these sanity checks to remain intact). Thus,
AssertionError is the next best thing.

It also doesn't extend the base Exception type, like LogicException
does, and thus there is now a clear distinction between Exception (which
are always runtime exceptions and should thus not be logic/assertion
exceptions anymore) and other Throwables that may involve both.