Add very barebones UseStatementInsertionCreator for class autocompletion
Currently this always inserts at the same location. It is now a matter
of completing the implementation by implementing the remaining tests,
code, as well as the sorting algorithm from the Atom autocompletion
package and the ability to determine whether separate newlines should
be added for use statements or not.

When finished, this will be used to generate TextEdit objects that can
be attached to class autocompletion suggestions. Applying these will
allow the client to add use statements for classes, just as currently
happens in the php-integrator autocompletion Atom package.

References #43