Replace fuse with levenshtein distance fuzzy matching again (performance)
fuse apparently has absolutely horrendous performance for fuzzy matching
with larger data sets. Fuzzy matching about 5000 array elements
(constant names) against a simple query takes around 150 milliseconds.
This is just for constant names and the time to do only the fuzzy
matching is already three times the target latency (which is about ~50
ms).

I thought fuzzy matching would be faster since fuse supports a threshold
where it stops evaluating the result. Contrary to what I expected,
levenshtein distance is many orders of magnitude faster and only takes
~10 ms for the same constants in a quick test. The difference might also
be related to the fact that levenshtein is part of PHP core and
implemented via C, whilst fuse is a PHP library with all the drawbacks
of being interpreted.

Unfortunately levenshtein is not as good as fuzzy matching at
actually... well... fuzzy matching. Hopefully I'll be able to mitigate
or improve it in the future.

Test parameters for fuse:

    $fuse = new \Fuse\Fuse(array_values($this->constantListProvider->getAll()), [
        // See also https://github.com/Loilo/Fuse#options
        'shouldSort'       => true,
        'caseSensitive'    => false,
        'threshold'        => 0.25,
        'maxPatternLength' => 128,
        'keys'             => ['name']
    ]);

    $fuse->search('PHP_EOL');

References #43