Perform approximate string matching before building suggestions (performance)
The previous approach was:
    1) Let all providers determine if they are relevant for the
       requested offset. If so, let them generate their unfiltered
       relevant, suggestions.
    2) Aggregate all suggestions
    3) Perform approximate string matching on all received suggestions,
       sort them by score.
    4) Take the 15 best suggestions.

The problem with this approach is that providers such as the constant
list provider usually generate over 5000 suggestions in a simple code
base (due to the sheer amount of them already available in PHP itself).
This is not even including classes, functions and namespaces from PHP
as well as the codebase itself.

It turns out that fetching the lists of constants themselves is
virtually instant, as it's already optimized by being managed by a
registry which is maintained across index updates. Just the simple
conversion of turning this data into AutocompletionSuggestion objects
takes around ~60 ms on a highend machine for ~5000 results, which is
absolutely unacceptable.

A solution to this problem is to filter out the irrelevant results
before even converting the data to AutocompletionSuggestion objects and
furthermore limiting the conversion to the x best results (i.e. around
15). This is one of the changes performed in this commit and it now
happens for constants, functions, classlikes as well as namespaces.

Another problem is that the scoring itself, currently using the
levenshtein algorithm, is not particularly slow (~6 ms for ~5000
results on the same hardware as before), but the sorting of this large
dataset by this score is.

The first solution to the sorting is the same as before: avoiding
sorting in the first place by filtering out as many results before the
sorting actually occurs. This was solved by letting the approximate
string matching determine a threshold above which results are simply
dropped (i.e. they look so differently from the input that they can
never be relevant).

With another performance hit down, it turns out that there are still
situations in which the threshold for the levenshtein distance is not
easily hit, such as when the user's input is one character or simply
empty. In these cases, we aren't able to filter out many suggestions and
fall back into the same issue where we are stuck with too many
suggestions.

This last problem is now optimized by partitioning the n suggestions
into m score partitions first. Afterwards, the partitions, where m <= n,
are sorted from best score to lowest score. This reduces the scope of
the sorting to partitions, rather than the suggestions, of which there
are usually considerably fewer (as suggestions often share the same
score). The best x results are then fetched from these partitions and
finally these results undergo a final sort (as they were just
partitioned, not sorted), which is limited to the desired amount of
items.

An overview of optimizations so far, based on the
testProvideAllFromStubs method in AutocompletionPerformanceTest:

  - Initial: ~800 ms
  - Avoid repeated locating of node at specific offset (fb84d46167311472c28626bf36d7a1d0eadfeee1): ~400 ms
  - Switch to levenshtein instead of fuse (2da2ce336541e1552902858df616515b3378ab26): ~250 ms
  - Drop irrelevant results early and optimize approximate string matching (this commit): ~60 ms

References #43