Don't return global autocompletion suggestions for empty prefix
This concerns classlikes, global functions, global constants, namespaces
and local variables. It is possible to return them, but it is often
useless for the client as the programmer has not even typed anything
yet, so there is nothing to complete.

Also, clients such as Atom are rather trigger happy when it comes to
autocompletion. When placing the cursor after a comma in a function
call using the mouse, autocompletion suggestions are immediately
retrieved. Returning classlikes here, such as before this commit, would
cause them to be included in this list. If the programmer then presses
the return key to create a newline, he confirms the suggestion he didn't
want instead; rather annoying behavior.

Member suggestions after i.e. $this are not affected by this, because
the user has taken explicit action to get there (i.e. by typing
"$this->" itself).

References #43