Allow more member suggestions than 15 results per provider
The main purpose of the limit was to avoid costly lists from generating
too many (less relevant) suggestions, such as classlikes, global
constants, global functions and namespaces. For suggestions such as
member methods and properties, this is less relevant since there usually
aren't that many of them - unless you have a huge class, which is
usually a code smell anyway.

This limits the large providers to 15 results each at most, like before.
The other providers themselves are unlimited, but the final aggregation
of the results from all providers is now limited to 50 results (where
this was also 15 before). As a result, more member suggestions will now
show up now.

References #43