Generate textEdit for classlike autocompletion suggestions
The insertText is also still generated, but the textEdit should get
precedence over the insertText, whenever available.

Some clients automatically determine the prefix to replace on their end
(e.g. Atom) and just paste the insertText we send back over this prefix.
This prefix sometimes differs from what we see as prefix as the namespace
separator (the backslash \) whilst these clients don't. Using a TextEdit
rather than a simple insertText ensures that the entire prefix is
replaced along with the insertion.

References #43