Fix position out of bounds error now that reindexing is disabled
We no longer reindex contents on every request, which caused the
"Position out of bounds" error to return because it was a
LogicException. The new version of the name qualification utilities
turn this exception into a runtime exception that results in the
request receiving an error response rather than the core crashing.

Contents should *probably* still be indexed before requests on the new
content are executed, but in real world scenarios this simply does not
appear to be realistic as most requests can't be answered without first
doing intensive processing on the input code. Such requests are usually
"read-only" requests such as autocompletion, which then all of a sudden
are also causing writes.

Also, in LSP, there is no concept of sending STDIN with these requests,
each change to a file is sent in separate requests. Also, even in that
scenario, read-only requests such as autocompletion will retain a higher
priority than reindex events triggered by these changes solely because
the user should not have to wait constantly for indexing to catch up
with his or her last typed characters, which usually bear no to little
significance on the results of the read-only request.

Fixes #169