Use LSPv3 $0 syntax to place cursor between parantheses after completion
This obsoletes placeCursorBetweenParentheses. The only use case this
breaks is that previously, if a function had no parameters and there
were no parantheses yet, the parantheses were inserted, but the cursor
was still placed after the insertion. After this change, they are still
placed inside the function call.

This does not occur that often in most code bases, though.

References #43