Fix large bulk indexing performance degradation due to cycle collection
Bulk indexing generates a very large amount of requests and responses
that need to be processes by the queue, such as the notification
responses that inform the client of indexing progress as well as one
request for every file that needs to be indexed and folder that needs
to be scanned.

Each and every one of these requests was triggering the
gc_collect_cycles call, even if there is nothing to collect. It was
added mostly as a safe guard until #168 can be fixed, so only perform
it every so often.