Replace AssertionError usage with LogicException
I was wrong in doing this previously. AssertionErrorS should be used
for assertions only and these are not assertions, which are meant to
e.g. maintain the internal consistency inside a class and the
interoperation of its own methods. For communicating errors (logic or
not, checked or unchecked) to other classes, assertions are not the
right answer.