Large refactor to use more objects (Position, TextDocumentItem, ...)
This replaces many uses of "int $offset" and "int $location" with
"Position" objects, introduces AutocompletionProviderContext on top of
it to get access to the prefix in every autocompletion provider, groups
type deduction parameters into a new TypeDeductionContext value object
and more.

Not all of the code base is using these objects yet, but I think we are
at least over half now; all the major components (signature help,
tooltips, autocompletion, indexing, ...) are now using them as inputs.
A few lower-level classes remain that don't, which will likely follow
in the future.

Using the Position object also ensures we are using character offsets
everywhere instead of byte offsets, which is needed for the LSP. It also
replaced some locations that were still only using lines instead of
proper offsets on that line as well.

References #111
References #113
References #217
References #213