Start rewriting and adding functionality for LSP support (WIP)
This commit contains all kinds of changes as the rewrite touches pretty
much everything of the command or request handling layer, as well as
how internal things are handled.

This is also very much a work-in-progress, so expect broken tests and
lots of problems.

Some things that have been changed (not an exhaustive list):

  - "initialize" was rewritten to be LSP-compliant
  - "exit" was rewritten to be LSP-compliant
  - "shutdown" was implemented
  - "initialized" notification is now handled and doesn't throw errors
  - "workspace/didChangeWatchedFiles" was implemented
  - "textDocument/didChange" was implemented
  - "autocomplete" was renamed to "textDocument/completion"
  - "tooltip" was renamed to "textDocument/hover"
  - "signatureHelp" was renamed to "textDocument/signatureHelp"
  - "reindex" was removed (no longer relevant)

The server also now requires a configuration file in
.serenata/config.json relative the root path of the project. The index
will also be saved in this folder and this will allow for additional
data to be stored there in the future. It supports pretty much the same
settings as are currently supported by the Atom package, but by having
it here it can be reused by all clients, not just Atom.

References #91