Keep track of document content for use in commands
We can't use STDIN to fetch document contents in requests anymore,
so we have to maintain a registry to access the most recent contents
of each document. We also can't load from files as the contents
inside the client editor may have been modified since the load from
disk.

References #91