Debounce indexing events to avoid them hogging the server
This is the same workaround the Atom package used to avoid
spamming the server with index requests on every buffer
change. Since for the LSP clients can send literally every
change to the server, it is now the server's responsibility
to do the debouncing.

Currently items are indexed after about 500 milliseconds,
which is the same default as was present in the Atom
client package.

References #91