Fix desynchronization of content in registry with actual content
Due to the didChange and didChangeWatchedFiles requests being marked
as low priority now (as they replaced the reindex request), other
requests such as the completion request got processed first. This
is a problem as the former request's handlers update the content of
the text document registry, which the latter request's handler needs
the state in the regsistry to be set to the very latest; otherwise
the passed position may be a position in a text buffer that doesn't
actually exist yet.

This solves the issue by reintroducing an internal index request
that carries the low priority instead. This way the didChange
and didChangeWatchedFiles requests can be handled with normal
priority. Clients such as Atom first send a didChange event
with the appropriate change before proceeding to send
requests for other functionality such as completion, so this
neatly processes the buffer update before proceeding to
process the completion request, which then carries a
position that points to the correct location and accompanies
the previously indexed content.

Additionally, this *may* just solve the problem that changes
caused by user interactions were not being prioritized over
other actions causing indexing requests, such as project
indexing. Project indexing now results in many smaller "index"
events via the demuxer, whilst user interactions will pass
through didChange and didChangeWatchedFiles, which are
debounced and then proceed to immediately index the files as
soon as the timer is hit. This should improve user experience
a bit.

Finally, this also fixes every file during project indexing
being debounced as well as it also passed through the
debouncingIndexer via de didChangeWatchedFiles request,
which was unnecessary overhead.

References #91