Relicense under the AGPLv3
This is a decision that I have been thinking about for quite some time
and I regret not having initially licensed Serenata (or PHP Integrator)
under the AGPL instead of the GPL.

Since both licenses are compatible ("Each place that the matrix states
GPLv3, the same statement about compatibility is true for AGPLv3 as 
well."
[1]) and I am also the copyright holder of probably 99%-ish of the
codebase and there are no large uses of Serenata that would experience 
any
adverse affects from this, this shouldn't pose any problem.

Also, I figured this is probably the time to still do it, as if Serenata
ever is to gain more contributions in the future, it will become
substantially harder then.

If anyone is wondering about the reasoning behind this: it is because, 
in
the last couple of years, there seems to be a move towards remote 
workspaces
and working on remote servers instead of locally in the editing and IDE 
scene
(LSP servers are inherently networked, Eclipse Ché, Xray, etc.), so I 
want
to make sure no one is able to "abuse" Serenata by making proprietary
modifications which benefit no one else in those scenarios due to the
application service provider loophole.

I'm happily investing time into this project for the benefit of myself 
and the
community, but I do not want anyone to be able just take this code and 
make
proprietary modifications for their own or company's personal or 
financial
gain, whilst not contributing those changes back to the community; I and
the contributors to Serenata already did all this work for free for you, 
it is
only fair that if you make modifications and want to spread them, you
contribute back so we might profit from them the same way you did from 
ours.

The source code of the previous versions will remain available under the
GPLv3 as was the case before, of course.

[1] https://www.gnu.org/licenses/gpl-faq.en.html#AllCompatibility