Fix crash with latest dependencies due to missing interface in Doctrine
Not sure how this happened, but it must have been inadvertently caused 
in one of our dependencies, since everything worked fine on the same PHP 
setup before updating the project's dependencies, so there must have 
been a change in behavior somewhere, which now executes this code and 
fails on it. 

Apparently Doctrine's AbstractManagerRegistry uses reflection to check 
if entities implement interfaces, but the interface that we passed did 
not exist, which is now a problem. Since it was probably always 
returning false before, just return an interface that entities will 
likely never implement (note that DateTimeInterface cannot be 
implemented by userland classes anyway).

Fixes #248