Index URIs configured in project instead of rootUri passed during init
The rootUri we receive is just the root of the folder that is loaded by 
the client, but it's not necessarily the location the user wants to 
index. This fixes that by explicitly indexing the URIs the user has 
configured, which will usually just be the same as the rootUri, but may 
include additional locations, such as stubs, remote locations, 
subfolders, or an entirely different location.

Fixes #243