Require that index database location is sent by client
This allows clients to determine the location of the index database 
entirely. As the location of the configuration file is also determined 
by the client, clients can now freely determine where they want to store 
these configurations and index databases for the user. Some clients may 
prefer to show a GUI instead of using an on-disk configuration file, or 
to store the index database in the /tmp folder, for example.

Implements #266