Fix no import generated when name with different kind is already present
If you try to import class A\B\Foo and a constant with name A\B\Foo is 
already imported, no import would be added. This is incorrect, because 
importing the same name multiple times is allowed in PHP, as long as the 
kind is different, i.e. you may import a constant, function and class 
with the same name simultaneously (but not two of the same kind and 
name, obviously).

Fixes #268