Fix incorrect class import location when only function imports exist
It was being placed in between the existing function imports instead of 
above them.

This only happened if you had a namespace present and didn't happen 
because it grouped the classlike import with the function imports, but 
rather because there were no existing classlike imports to use as a 
reference point, so it thought it had to insert them on the fallback 
location, which then turned out to be incorrect.

Fixes #267