Move most completion FQCNs to detail instead of label
This moves the FQCN (or FQSEN) for classlikes, functions and constants 
to the detail property of completion suggestions. This keeps the labels 
themselves much shorter, only identifying the most important part of 
these elements: their name.

The detail can then serve to distinguish items when multiple items exist 
with the same name. Whilst this happens often enough to be relevant 
("Request" appears to be used a lot), the majority of suggestions will 
be alone in their existence, in which case it will be visually much less 
overwhelming if we can omit the prefix of the name there.

Fixes #269